pyproject.toml
[tool.poetry]
name = "datamodel-deploy-template"
version = "0.1.0"
description = "Deployment template for data models."
readme = "README.md"
authors = ["Cognite <support@cognite.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.10"
cognite-sdk = {version = ">=6.29.0"}
pyyaml = "^6"


[build-system]
requires = ["poetry-core>=1.3"]
build-backend = "poetry.core.masonry.api"
